feat: add weakness overview screen
